requirements_dev.txt
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
black==20.8b1
cairo-helper==0.0.1
cfgv==3.2.0
click==7.1.2
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
identify==1.5.0
image-helper==0.1.0
iniconfig==1.0.1
isort==5.5.2
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.4
parso==0.7.1
pathspec==0.8.0
pluggy==0.13.1
pre-commit==2.7.1
py==1.9.0
pycairo==1.19.1
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
python-jsonrpc-server==0.4.0
python-language-server==0.35.0
PyYAML==5.3.1
regex==2020.7.14
rope==0.17.0
six==1.15.0
snowballstemmer==2.0.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==3.2.0
virtualenv==20.0.31
wrapt==1.12.1
yapf==0.30.0
